+++ internal-staff-directory.php
@@ -22,5 +22,6 @@ define( 'EMPLOYEE_DIR_PLUGIN_DIR',  plugin_dir_path( __FILE__ ) );
 define( 'EMPLOYEE_DIR_PLUGIN_URL',  plugin_dir_url( __FILE__ ) );
 
 require_once EMPLOYEE_DIR_PLUGIN_DIR . 'includes/profile.php';
+require_once EMPLOYEE_DIR_PLUGIN_DIR . 'includes/settings.php';
 require_once EMPLOYEE_DIR_PLUGIN_DIR . 'includes/directory.php';
 require_once EMPLOYEE_DIR_PLUGIN_DIR . 'includes/admin.php';